backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
alembic
langchain
langgraph
langchain-groq
pandas
statsmodels
pytest
httpx
python-jose
passlib
python-dateutil


